+++ doc/release/release_notes.rst
@@ -9,6 +9,12 @@ Release Notes
 
 .. release:: Upcoming
 
+    .. change:: new
+        :tags: Session
+
+        Session now supports an `ensure` method that will if necessary update or
+        create an entity with the given data.
+
     .. change:: fixed
         :tags: Event Hub