+++ doc/release/release_notes.rst
@@ -15,6 +15,10 @@ Release Notes
         Session now supports an `ensure` method that will if necessary update or
         create an entity with the given data.
 
+        .. seealso::
+
+            :ref:`api_reference/session/ensure`
+
     .. change:: fixed
         :tags: Event Hub